% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxNd.R
\name{approxNd}
\alias{approxNd}
\title{approxNd}
\usage{
approxNd(mat, list_seq, list_q)
}
\arguments{
\item{mat}{Array on N dimensions}

\item{list_seq}{Boundary seq on N dimensions}

\item{list_q}{Query values on N dimensions}
}
\value{
A numeric value
}
\description{
approxNd
}
\examples{
list_seq = list(1:2, 3:4, 5:6)
list_q   = list(1.5, 3.5, 5.5)
mat = array(runif(8),
            dim = lapply(list_seq, length),
            dimnames = list_seq)
approxNd(mat, list_seq, list_q)
# the interpolated value should be identical to the mean values
mean(mat)

list_seq = list(1:2, 3:4, 5:6, 7:8, 9:10, 11:12)
list_q   = list(1.5, 3.5, 5.5, 7.5, 9.5, 11.5)
mat = array(runif(2^length(list_seq)),
            dim = lapply(list_seq, length),
            dimnames = list_seq)
approxNd(mat, list_seq, list_q)
mean(mat)

}
