% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_spec}
\alias{demo_spec}
\title{demo_spec}
\format{
\subsection{\code{data.table}}{

A data table with 202 rows and 14 columns:
\describe{
\item{SampleID}{SampleID}
\item{wavelen}{Wavelength}
\item{Theta}{Viewing zenith angle}
\item{Phi}{Relative azimuth angle}
\item{Rrs}{Remote sensing reflectance (above-water) from Theta and Phi direction}
\item{Rrs}{Remote sensing ratio (underwater) on the nadir direction}
\item{suntheta}{Solar zenith angle}
\item{windspd}{Wind speed}
\item{cloud}{cloud coverage}
\item{at}{Total absorption}
\item{bbt}{Total backward scattering}
\item{bbw}{Backward scattering for pure water}
\item{Temp}{Water temperautre}
\item{Sal}{Water salinity}
}
}
}
\usage{
demo_spec
}
\description{
demo_spec
}
\note{
This is a demo spectra used to test the model.
}
\keyword{datasets}
