% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RrsTrans-package.R
\docType{package}
\name{RrsTrans-package}
\alias{RrsTrans}
\alias{RrsTrans-package}
\title{RrsTrans: Transfer remote-sensing ratio (rrs) to remote-sensing reflectance (Rrs)}
\description{
Transfer models to determine the above-water remote-sensing reflectance from the underwater remote-sensing ratio. The package is an appendix of our recent study entitled "A transfer model to determine the above-water remote-sensing reflectance from the underwater remote-sensing ratio" by Bi et al. (202x). Please refer to this paper for more details.
}
\author{
\strong{Maintainer}: Shun Bi \email{Shun.Bi@hereon.de} (\href{https://orcid.org/0000-0002-2668-5808}{ORCID})

}
\keyword{internal}
