% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_funcs.R
\name{trans_function}
\alias{trans_function}
\alias{trans_Lee02}
\alias{trans_Mobley99}
\alias{trans_Loisel08}
\alias{trans_Lee11}
\alias{trans_Bi22}
\alias{trans_Morel96}
\title{Transfer models from underwater remote-sensing ratio (rrs) to
above-water remote-sensing reflectance (Rrs)}
\usage{
trans_Lee02(rrs)

trans_Mobley99(rrs)

trans_Loisel08(rrs, suntheta)

trans_Lee11(at, bbt, bbw, Theta, Phi, suntheta)

trans_Bi22(rrs, Theta, Phi, suntheta, windspd, cloud = NULL, nw)

trans_Morel96(rrs, wavelen, Theta, suntheta)
}
\arguments{
\item{rrs}{remote sensing reflectance just beneath the water surface}

\item{suntheta}{Solar zenith angle in deg}

\item{at}{total absorption in m-1}

\item{bbt}{total backward scattering in m-1}

\item{bbw}{water backward scattering in m-1}

\item{Theta}{Viewing zenith angle in deg}

\item{Phi}{Relative azimuth angle in deg}

\item{windspd}{Wind speed in meter/s default as 5 m/s}

\item{cloud}{Cloud fraction default as 0 (a clear sky) and 1 is a overcast
sky}

\item{nw}{Refractive index of water}

\item{wavelen}{wavelength in nm}
}
\description{
Transfer models from underwater remote-sensing ratio (rrs) to
above-water remote-sensing reflectance (Rrs)
}
\examples{

# parameters preparation
wavelen  = c(410, 440, 560, 620, 670, 800)
rrs      = c(0.00557, 0.00834, 0.0208, 0.01035, 0.00624, 0.00132)
Rrs      = c(0.00297, 0.0045, 0.01158, 0.00566, 0.00339, 0.00072)
at       = c(0.7576, 0.5058, 0.1964, 0.3547, 0.5321, 2.2494)
bbt      = c(0.04685, 0.04536, 0.04089, 0.03851, 0.0364, 0.03578)
bbw      = c(0.00306, 0.00226, 0.00082, 0.00054, 0.00039, 0.00019)
suntheta = 59.90437
Theta    = 40
Phi      = 135
windspd  = 6
cloud    = 0
Temp     = 1.1
Sal      = 31.1

# run models
nw = WOPP(Temp, Sal, wavelen)$nw
Rrs_Lee02 = trans_Lee02(rrs)
Rrs_Loisel08 = trans_Loisel08(rrs, suntheta)
Rrs_Lee11 = trans_Lee11(at, bbt, bbw, Theta, Phi, suntheta)
Rrs_Bi22 = trans_Bi22(rrs, Theta, Phi, suntheta, windspd, cloud, nw = nw)
Rrs_Mobley99 = trans_Mobley99(rrs)
Rrs_Morel96 = trans_Morel96(rrs, wavelen, Theta, suntheta)

# plot
pal_col = rainbow(n=6)
plot(wavelen, Rrs, col = "black", type = "l",
     xlim = c(400, 810), ylim = c(0, 0.012),
     xlab = "Wavelength [nm]", ylab = "Rrs [1/sr]")
lines(wavelen, Rrs_Bi22, col = pal_col[1], type = "l")
lines(wavelen, Rrs_Lee02, col = pal_col[2], type = "l")
lines(wavelen, Rrs_Loisel08, col = pal_col[3], type = "l")
lines(wavelen, Rrs_Lee11, col = pal_col[4], type = "l")
lines(wavelen, Rrs_Morel96, col = pal_col[5], type = "l")
lines(wavelen, Rrs_Mobley99, col = pal_col[6], type = "l")
lines(wavelen, Rrs, col = "black", type = "l", lty = 2)
legend(x = 810, y = 0.012, xjust = 1, yjust = 1,
       c("HydroLight", "Bi22", "Lee02", "Loisel08",
         "Lee11", "Morel96", "Mobley99"),
       col = c("black", pal_col), lty = c(1, 2, rep(1, 5)),
       x.intersp = 0.5, y.intersp = 0.8)

}
\references{
\itemize{
\item Lee, ZhongPing, Kendall L. Carder, and Robert A. Arnone. “Deriving Inherent
Optical Properties from Water Color: A Multiband Quasi-Analytical Algorithm
for Optically Deep Waters.” Applied Optics 41, no. 27 (September 20, 2002):
5755. https://doi.org/10.1364/AO.41.005755.
\item Lee, Zhongping, Keping Du, Kenneth J. Voss, Giuseppe Zibordi, Bertrand
Lubac, Robert Arnone, and Alan Weidemann. “An
Inherent-Optical-Property-Centered Approach to Correct the Angular Effects in
Water-Leaving Radiance.” Applied Optics 50, no. 19 (July 1, 2011): 3155.
https://doi.org/10.1364/AO.50.003155.
\item Loisel, Hubert. “Rrs(0+) -> Rrs(0-) & Water Coefficients.” Ocean Color
Forum, Inherent OPtical Properties Workshop, 2008.
http://oceancolor.gsfc.nasa.gov/forum/oceancolor/topic_show.pl?tid=2657.
\item Mobley, Curtis D. “Estimation of the Remote-Sensing Reflectance from
above-Surface Measurements.” Applied Optics 38, no. 36 (December 20, 1999):
7442. https://doi.org/10.1364/AO.38.007442.
\item Morel, André, and Bernard Gentili. “Diffuse Reflectance of Oceanic Waters
III Implication of Bidirectionality for the Remote-Sensing Problem.” Applied
Optics 35, no. 24 (August 20, 1996): 4850. https://doi.org/10.1364/AO.35.004850.
}
}
